+++ src/views/datetime.blade.php
@@ -1,8 +1,12 @@
 <div class="control-group {{ $field->containerClass() }}">
-    <label class="control-label" for="{{$field->id}}">{{$field->label}}</label>
+    <label class="control-label" data-format="dd/MM/yyyy hh:mm:ss" for="{{$field->id}}">{{$field->label}}</label>
     <div class="controls">
-        <input class="input-medium" id="{{$field->id}}" name="{{$field->name}}[date]" type="date" value="{{$field->value['date']}}"/>
-        <input class="input-mini"   id="{{$field->id}}" name="{{$field->name}}[time]" type="time" value="{{$field->value['time']}}"/>
+        <div class="input-append datetime">
+            <input data-format="dd/MM/yyyy hh:mm:ss" {{ $field->attributes() }}>
+            <span class="add-on">
+                <i data-time-icon="icon-time" data-date-icon="icon-calendar"></i>
+            </span>
+        </div>
         {{ $field->help() }}
         {{ $field->errors() }}
     </div>